refactor: EasyCodef 관련 추상 클래스 분리 완료
